refactor: move suggestions to their own package
