+++ internal/suggestions/suggestions_test.go
@@ -1,4 +1,4 @@
-package firehose
+package suggestions
 
 import (
 	"encoding/json"
@@ -8,15 +8,16 @@ import (
 	"time"
 
 	"arabica/internal/atproto"
+	"arabica/internal/firehose"
 
 	"github.com/stretchr/testify/assert"
 )
 
-func newTestFeedIndex(t *testing.T) *FeedIndex {
+func newTestFeedIndex(t *testing.T) *firehose.FeedIndex {
 	t.Helper()
 	dir := t.TempDir()
 	path := filepath.Join(dir, "test-index.db")
-	idx, err := NewFeedIndex(path, 1*time.Hour)
+	idx, err := firehose.NewFeedIndex(path, 1*time.Hour)
 	assert.NoError(t, err)
 	t.Cleanup(func() {
 		idx.Close()
@@ -25,7 +26,7 @@ func newTestFeedIndex(t *testing.T) *FeedIndex {
 	return idx
 }
 
-func insertRecord(t *testing.T, idx *FeedIndex, did, collection, rkey string, fields map[string]interface{}) {
+func insertRecord(t *testing.T, idx *firehose.FeedIndex, did, collection, rkey string, fields map[string]interface{}) {
 	t.Helper()
 	fields["$type"] = collection
 	fields["createdAt"] = time.Now().Format(time.RFC3339)
@@ -35,7 +36,7 @@ func insertRecord(t *testing.T, idx *FeedIndex, did, collection, rkey string, fi
 	assert.NoError(t, err)
 }
 
-func TestSearchEntitySuggestions_PrefixMatch(t *testing.T) {
+func TestSearch_PrefixMatch(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
 	insertRecord(t, idx, "did:plc:alice", atproto.NSIDRoaster, "r1", map[string]interface{}{
@@ -47,7 +48,7 @@ func TestSearchEntitySuggestions_PrefixMatch(t *testing.T) {
 		"location": "Oakland, CA",
 	})
 
-	results, err := idx.SearchEntitySuggestions(atproto.NSIDRoaster, "bl", 10)
+	results, err := Search(idx, atproto.NSIDRoaster, "bl", 10)
 	assert.NoError(t, err)
 	assert.Len(t, results, 2)
 	// Both match prefix "bl"
@@ -55,20 +56,20 @@ func TestSearchEntitySuggestions_PrefixMatch(t *testing.T) {
 	assert.Equal(t, "Blue Bottle", results[1].Name)
 }
 
-func TestSearchEntitySuggestions_CaseInsensitive(t *testing.T) {
+func TestSearch_CaseInsensitive(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
 	insertRecord(t, idx, "did:plc:alice", atproto.NSIDRoaster, "r1", map[string]interface{}{
 		"name": "Stumptown Coffee",
 	})
 
-	results, err := idx.SearchEntitySuggestions(atproto.NSIDRoaster, "STUMP", 10)
+	results, err := Search(idx, atproto.NSIDRoaster, "STUMP", 10)
 	assert.NoError(t, err)
 	assert.Len(t, results, 1)
 	assert.Equal(t, "Stumptown Coffee", results[0].Name)
 }
 
-func TestSearchEntitySuggestions_SubstringMatch(t *testing.T) {
+func TestSearch_SubstringMatch(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
 	insertRecord(t, idx, "did:plc:alice", atproto.NSIDRoaster, "r1", map[string]interface{}{
@@ -77,13 +78,13 @@ func TestSearchEntitySuggestions_SubstringMatch(t *testing.T) {
 	})
 
 	// Search by location substring
-	results, err := idx.SearchEntitySuggestions(atproto.NSIDRoaster, "floyd", 10)
+	results, err := Search(idx, atproto.NSIDRoaster, "floyd", 10)
 	assert.NoError(t, err)
 	assert.Len(t, results, 1)
 	assert.Equal(t, "Red Rooster Coffee", results[0].Name)
 }
 
-func TestSearchEntitySuggestions_Deduplication(t *testing.T) {
+func TestSearch_Deduplication(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
 	// Two users have the same roaster (different case/whitespace)
@@ -96,7 +97,7 @@ func TestSearchEntitySuggestions_Deduplication(t *testing.T) {
 		"name": "Counter Culture",
 	})
 
-	results, err := idx.SearchEntitySuggestions(atproto.NSIDRoaster, "counter", 10)
+	results, err := Search(idx, atproto.NSIDRoaster, "counter", 10)
 	assert.NoError(t, err)
 	assert.Len(t, results, 1)
 	assert.Equal(t, 2, results[0].Count)
@@ -104,7 +105,7 @@ func TestSearchEntitySuggestions_Deduplication(t *testing.T) {
 	assert.Equal(t, "Durham, NC", results[0].Fields["location"])
 }
 
-func TestSearchEntitySuggestions_Limit(t *testing.T) {
+func TestSearch_Limit(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
 	for i := 0; i < 5; i++ {
@@ -115,12 +116,12 @@ func TestSearchEntitySuggestions_Limit(t *testing.T) {
 		})
 	}
 
-	results, err := idx.SearchEntitySuggestions(atproto.NSIDGrinder, "grinder", 3)
+	results, err := Search(idx, atproto.NSIDGrinder, "grinder", 3)
 	assert.NoError(t, err)
 	assert.Len(t, results, 3)
 }
 
-func TestSearchEntitySuggestions_ShortQuery(t *testing.T) {
+func TestSearch_ShortQuery(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
 	insertRecord(t, idx, "did:plc:alice", atproto.NSIDRoaster, "r1", map[string]interface{}{
@@ -128,33 +129,33 @@ func TestSearchEntitySuggestions_ShortQuery(t *testing.T) {
 	})
 
 	// Query too short (< 2 chars)
-	results, err := idx.SearchEntitySuggestions(atproto.NSIDRoaster, "a", 10)
+	results, err := Search(idx, atproto.NSIDRoaster, "a", 10)
 	assert.NoError(t, err)
 	assert.Empty(t, results)
 
 	// 2 chars should work
-	results, err = idx.SearchEntitySuggestions(atproto.NSIDRoaster, "ab", 10)
+	results, err = Search(idx, atproto.NSIDRoaster, "ab", 10)
 	assert.NoError(t, err)
 	assert.Len(t, results, 1)
 }
 
-func TestSearchEntitySuggestions_EmptyQuery(t *testing.T) {
+func TestSearch_EmptyQuery(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
-	results, err := idx.SearchEntitySuggestions(atproto.NSIDRoaster, "", 10)
+	results, err := Search(idx, atproto.NSIDRoaster, "", 10)
 	assert.NoError(t, err)
 	assert.Empty(t, results)
 }
 
-func TestSearchEntitySuggestions_UnknownCollection(t *testing.T) {
+func TestSearch_UnknownCollection(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
-	results, err := idx.SearchEntitySuggestions("unknown.collection", "test", 10)
+	results, err := Search(idx, "unknown.collection", "test", 10)
 	assert.NoError(t, err)
 	assert.Empty(t, results)
 }
 
-func TestSearchEntitySuggestions_GrinderFields(t *testing.T) {
+func TestSearch_GrinderFields(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
 	insertRecord(t, idx, "did:plc:alice", atproto.NSIDGrinder, "g1", map[string]interface{}{
@@ -163,14 +164,14 @@ func TestSearchEntitySuggestions_GrinderFields(t *testing.T) {
 		"burrType":    "conical",
 	})
 
-	results, err := idx.SearchEntitySuggestions(atproto.NSIDGrinder, "1zp", 10)
+	results, err := Search(idx, atproto.NSIDGrinder, "1zp", 10)
 	assert.NoError(t, err)
 	assert.Len(t, results, 1)
 	assert.Equal(t, "hand", results[0].Fields["grinderType"])
 	assert.Equal(t, "conical", results[0].Fields["burrType"])
 }
 
-func TestSearchEntitySuggestions_BeanFields(t *testing.T) {
+func TestSearch_BeanFields(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
 	insertRecord(t, idx, "did:plc:alice", atproto.NSIDBean, "b1", map[string]interface{}{
@@ -181,14 +182,14 @@ func TestSearchEntitySuggestions_BeanFields(t *testing.T) {
 	})
 
 	// Search by origin
-	results, err := idx.SearchEntitySuggestions(atproto.NSIDBean, "ethiopia", 10)
+	results, err := Search(idx, atproto.NSIDBean, "ethiopia", 10)
 	assert.NoError(t, err)
 	assert.Len(t, results, 1)
 	assert.Equal(t, "Ethiopian Yirgacheffe", results[0].Name)
 	assert.Equal(t, "Light", results[0].Fields["roastLevel"])
 }
 
-func TestSearchEntitySuggestions_BrewerFields(t *testing.T) {
+func TestSearch_BrewerFields(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
 	insertRecord(t, idx, "did:plc:alice", atproto.NSIDBrewer, "br1", map[string]interface{}{
@@ -196,13 +197,13 @@ func TestSearchEntitySuggestions_BrewerFields(t *testing.T) {
 		"brewerType": "Pour-Over",
 	})
 
-	results, err := idx.SearchEntitySuggestions(atproto.NSIDBrewer, "hario", 10)
+	results, err := Search(idx, atproto.NSIDBrewer, "hario", 10)
 	assert.NoError(t, err)
 	assert.Len(t, results, 1)
 	assert.Equal(t, "Pour-Over", results[0].Fields["brewerType"])
 }
 
-func TestSearchEntitySuggestions_SortOrder(t *testing.T) {
+func TestSearch_SortOrder(t *testing.T) {
 	idx := newTestFeedIndex(t)
 
 	// "Alpha Roasters" used by 3 people
@@ -213,7 +214,7 @@ func TestSearchEntitySuggestions_SortOrder(t *testing.T) {
 	// "Alpha Beta" used by 1 person
 	insertRecord(t, idx, "did:plc:dave", atproto.NSIDRoaster, "r4", map[string]interface{}{"name": "Alpha Beta"})
 
-	results, err := idx.SearchEntitySuggestions(atproto.NSIDRoaster, "alpha", 10)
+	results, err := Search(idx, atproto.NSIDRoaster, "alpha", 10)
 	assert.NoError(t, err)
 	assert.Len(t, results, 2)
 	// More popular first
